backend/services/ai-risk-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
PyJWT==2.10.1
numpy==2.3.2
pandas==2.3.1
scikit-learn==1.7.1
joblib==1.5.1
httpx==0.28.1
PyYAML==6.0.2
matplotlib==3.10.5
seaborn==0.13.2
pytest==8.4.1
